backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
boto3==1.35.85
python-dotenv==1.0.1
